Changing install directories to match with the default prefix of cmake.
This will allow cmake to find the config modules without adjusting the path.

The rest of the code needs to be adjusted to use the new installation directories.

Essentially:

find(TivaWare) will bring in tivaware
find(TivaStartup) will bring in the startup library
find(TivaCMake)  will put all the auxillary cmake modules on the module path