Modifying the way paths are found.
There are two types of cmake packages:
Modules (FindYYY.cmake) and Config YYYConfig.cmake

3rd party Config packages can be placed in standard locations on the system
but 3rd party Modules are only searched for in the base cmake install directory.

CMAKE_MODULE_PATHS and CMAKE_PREFIX_PATHS control searching for Modules and COnfig packages
respectively.

tiva-cmake uses both kinds of modules and they may be in different locations
depending on if the user is compiling code using this as a template or has installed
tiva-cmake on their system.

The way I handle paths is as follows:
1. TivaCMakeConfig.cmake sets the module path to point to the tiva-cmake modules.
2. All code outside of src/ uses find_package(... HINTS) to specify the location
   of the tiva-cmake/cmake directory.
   If the project is installed systemwide, the config packages are foudn in the usual places
   otherwise they are found according to the hint.
3. src/preload.cmake finds the TivaCMakeCOnfig package, thus setting everything up.
   It also sets up search paths for a default toolchain file.
4.  The call to find_package(TivaCMake) in src/TivaCMake is unnecessary since it is in pre-load,
 but technically the project works without a preload