Automatically including the startup code with add_executable
The reason for this is that there are problems with transiently depending on an OBJECT library
in different projects and also every tiva executable needs startup code or it won't run.
Thus, I consider the startup code more as part of the standard library than a library you
would want to optionally link against.

This behavior can be disabled by setting TIVA_CMAKE_USE_CUSTOM_STARTUP to true,
allowing the use of custom startup code instead of the code provided by tiva-cmake.