Fixed a bug related to overriding add_executable.
The add_executable version that I implemented did not work if somebody
created an IMPORTED target.  The built-in git module does this so
tiva-cmake could not be used in conjunction with, for example cmakeme.

Now, I explicitly determine if the user is attempting to create an IMPORTED executable
and then don't run the offending code.

I am unsure of how sustainable this is.  There could be other edge-cases where
my method of overriding add_executable does not work.