I changed the gcc startup code to call main directly.
Before it was calling _start(), which is the newlib startup code.
However, this was causing hard faults, I think because I was doing something
redundant with that startup code in terms of setting up the stack.

I am goign to explore if I can eliminate this custom startup code entirely and replace it with a call to start.