Added stub functions to suppress warnings that various functions were stubs
It seems like these warnings were added in a way such that if the stub from nosys was ever
linked, it would produce a warning

Basically, the syscall will fail if any code calls it.
There is (at some low level) some code in libg.a that calls some of these stubs.

Our code never called anything that uses them, but I can't seem to remove the need for it.
So instaed I wrote our own stubs to suppress the warning.

See
https://stackoverflow.com/questions/73742774/gcc-arm-none-eabi-11-3-is-not-implemented-and-will-always-fail
(accessed on 12/15/2025)